refactor(chat): remove legacy rag retrieval path
