+++ docs/codex-review-map.md
@@ -21,8 +21,9 @@ Clawpatch owns the semantic feature map and explicit finding/fix loop. Codex sti
 | Quiz and lead capture | `src/app/quiz/`, `src/components/quiz/`, `src/lib/quiz/` | German copy, field normalization, answered-vs-skipped semantics, lead lifecycle | `npm run test:node`, relevant `tests/quiz-*.test.ts`, `tests/quiz-onboarding-e2e.spec.ts` |
 | Onboarding and profile shaping | `src/app/onboarding/`, `src/components/onboarding/`, `src/lib/onboarding/`, `src/lib/profile/`, `src/hooks/use-hair-profile.ts` | `hair_texture` vs `thickness`, persistence safety, none-state handling, profile-derived signals | `tests/onboarding-*.test.ts`, `tests/hair-profile-derived.test.ts`, `tests/profile-*.spec.ts` |
 | Recommendation engine | `src/lib/recommendation-engine/`, `src/lib/product-specs/`, category constants and backfill scripts | Deterministic mappings, conservative fallbacks, `recommendation_meta`, category-specific fit logic | `tests/recommendation-engine-*.test.ts`, category flow specs, `npm run test:contracts` |
-| Agentic chat and tools | `src/app/api/chat/`, `src/lib/agent/`, `src/components/chat/`, `src/hooks/use-chat.ts` | Tool routing, product selection, answer composition, visible failure handling, German response UX | `npm run test:agent`, `npm run test:chat:ci`, selected Playwright chat specs |
-| RAG and memory | `src/lib/rag/`, `src/app/api/memory/`, `scripts/eval-retrieval.ts` | Retrieval thresholds, memory extraction, clarification state, source grounding | `npm run test:retrieval:ci`, `tests/conversation-state.spec.ts`, `tests/user-memory.spec.ts` |
+| Agentic chat and tools | `src/app/api/chat/`, `src/lib/agent/`, `src/lib/agent-v2/`, `src/components/chat/`, `src/hooks/use-chat.ts` | Tool routing, product selection, answer composition, visible failure handling, German response UX | `npm run test:agent`, `npm run test:chat:ci`, selected Playwright chat specs |
+| Chat memory, state, and traces | `src/lib/chat-runtime/`, `src/app/api/memory/`, `src/app/api/chat/`, chat trace tests | Memory extraction, conversation state, debug trace shape, source/projection grounding | `tests/conversation-state.spec.ts`, `tests/user-memory.spec.ts`, `tests/chat-debug-trace.spec.ts` |
+| Product matching and catalog chunks | `src/lib/product-matching/`, `scripts/ingest-product-chunks.ts`, `scripts/eval-retrieval.ts` | Catalog chunk shape, product matching determinism, retrieval eval ingestion boundaries | `npm run test:retrieval:ci`, `tests/product-matcher.spec.ts`, `tests/product-list-chunks.test.ts` |
 | Langfuse and eval loop | `src/lib/langfuse/`, `src/lib/openai/`, `scripts/eval-chat/`, `scripts/langfuse/` | Trace masking, prompt fallback behavior, dataset metadata, score publication | `tests/langfuse-*.test.ts`, `npm run test:chat:judge`, `npm run test:chat:langfuse` when comparing runs |
 | Stripe, auth, and access gates | `src/app/api/stripe/`, `src/lib/stripe/`, `src/app/auth/`, `src/lib/auth/`, `src/proxy.ts` | Checkout activation, webhook idempotency, session routing, pricing gates, password recovery | `tests/stripe-*.spec.ts`, `tests/auth-*.test.ts`, `tests/checkout-activation.spec.ts` |
 | Supabase schema and policies | `supabase/migrations/`, `src/lib/supabase/`, data ingestion/backfill scripts | RLS assumptions, admin boundaries, enum drift, migration reversibility, local-secret handling | Targeted migration review, affected persistence tests, `npm run test:node` |
@@ -63,7 +64,8 @@ Clawpatch-generated test links are useful triage hints, not proof of adequate co
 | Onboarding and profile shaping | 19 | 6 library, 1 route, 12 UI flow | 1 | `feat_library_0153fe859f` source group `src/components/onboarding/screens`; `feat_route_e82b77cd40` Route `/onboarding` |
 | Recommendation engine | 7 | 7 library | 0 | `feat_library_d9d76394b4` source group `src/lib/recommendation-engine/categories`; `feat_library_a72310a0b6` source group `src/lib/recommendation-engine` |
 | Agentic chat and tools | 22 | 9 library, 3 route, 10 UI flow | 3 | `feat_route_4f2453eb75` Route `/api/chat`; `feat_route_4c11dbaf4b` Route `/api/chat/feedback`; `feat_library_82fa133a7f` source group `src/lib/agent/orchestrator` |
-| RAG and memory | 11 | 9 library, 2 route | 5 | `feat_library_26b3ce8594` source group `src/lib/rag/retrieval`; `feat_route_4ccd39b375` Route `/api/memory`; `feat_route_6355863e58` Route `/api/memory/:id` |
+| Chat memory, state, and traces | 11 | 9 library, 2 route | 5 | Current live review focus is `src/lib/chat-runtime`, `/api/chat`, and `/api/memory`; generated inventory counts will be refreshed in Task 6 |
+| Product matching and catalog chunks | 0 | Pending refreshed Clawpatch map | 0 | Current live review focus is `src/lib/product-matching`, product chunk ingestion, and retrieval eval scripts until Task 6 refreshes the generated inventory |
 | Langfuse and eval loop | 4 | 3 library, 1 service | 0 | `feat_library_f4f86a7094` source group `src/lib/langfuse`; `feat_library_2e94e5b193` source group `scripts/langfuse`; `feat_library_77bde163e7` source group `scripts/eval-chat` |
 | Stripe, auth, and access gates | 12 | 5 library, 7 route | 8 | `feat_route_a92aa7d78f` Route `/api/stripe/webhook`; `feat_route_7909897458` Route `/api/stripe/create-checkout-session`; `feat_route_717e4d949f` Route `/api/auth/send-magic-link` |
 | Supabase schema and policies | 1 | 1 library | 0 | `feat_library_fd97feb5a0` source group `src/lib/supabase`; migrations are covered by the committed slice map rather than rich Clawpatch records today |
