+++ package.json
@@ -35,7 +35,7 @@
     "test:chat:langfuse": "tsx scripts/eval-chat/run.ts --langfuse-publish",
     "test:node": "tsx --test tests/*.test.ts tests/*.test.tsx",
     "test:agent": "tsx --test tests/agent-guidance.spec.ts tests/agent-guidance-tracing.spec.ts tests/agentic-tool-loop.spec.ts tests/agent-route-packet.spec.ts tests/agent-get-user-context.spec.ts tests/agent-select-products-tool.spec.ts tests/agent-routine-tool.spec.ts tests/agent-shadow.spec.ts tests/agent-v2-production-chat-pipeline.spec.ts tests/agent-final-render-prompt.spec.ts tests/agent-compare-api.spec.ts tests/agent-compare-product-trace.spec.ts tests/agent-compare-runner.spec.ts tests/agent-compare-test-users.spec.ts tests/agent-v2-contracts.spec.ts tests/agent-v2-guidance-compiler.spec.ts tests/agent-v2-tool-projections.spec.ts tests/agent-v2-final-answer-validator.spec.ts tests/agent-v2-responses-runtime.spec.ts tests/agent-v2-compare-runner.spec.ts tests/agent-v2-manual-regression.spec.ts",
-    "test:playwright:contracts": "playwright test tests/chat-debug-trace.spec.ts tests/conditioner-reranker.spec.ts tests/conversation-state.spec.ts tests/leave-in-decision.spec.ts tests/mask-flow.spec.ts tests/routine-planner.spec.ts tests/shampoo-flow.spec.ts tests/stripe-gating.spec.ts tests/stripe-intervals.spec.ts tests/stripe-webhook-handlers.spec.ts tests/user-memory.spec.ts --project=chromium",
+    "test:playwright:contracts": "playwright test tests/chat-debug-trace.spec.ts tests/conversation-state.spec.ts tests/stripe-gating.spec.ts tests/stripe-intervals.spec.ts tests/stripe-webhook-handlers.spec.ts tests/user-memory.spec.ts --project=chromium",
     "test:contracts": "npm run test:node && npm run test:agent && npm run test:playwright:contracts",
     "test:chat:ci": "tsx scripts/eval-chat/run.ts --skip-judge --ci-smoke",
     "test:retrieval:ci": "tsx scripts/eval-retrieval.ts --hybrid-only --min-hybrid-ndcg10 0.72 --min-hybrid-recall20 0.88 --skip-unannotated-gold-set",
