--- src/lib/rag/retriever.ts
@@ -1,114 +1,0 @@
-import type { ContentChunk, HairProfile, IntentType } from "@/lib/types"
-import {
-  COMMUNITY_QA_PRODUCT_BOOST,
-  SHAMPOO_CONCERN_MATCH_BOOST,
-  THICKNESS_MATCH_BOOST,
-} from "@/lib/rag/retrieval-constants"
-
-export interface RetrievedChunk extends ContentChunk {
-  similarity: number
-  weighted_similarity: number
-  /** Which retrieval path produced this chunk */
-  retrieval_path?: "dense" | "lexical" | "hybrid"
-  /** Raw dense cosine similarity score */
-  dense_score?: number
-  /** Raw lexical BM25 rank score */
-  lexical_score?: number
-  /** RRF fused score (only present for hybrid path) */
-  fused_score?: number
-}
-
-export interface RetrieveOptions {
-  intent?: IntentType
-  hairProfile?: HairProfile | null
-  metadataFilter?: Record<string, string>
-  shampooConcern?: string | null
-  count?: number
-  /** Subqueries generated by decomposition. Retained for legacy trace compatibility. */
-  subqueries?: string[]
-  /** Legacy candidate count option. No longer used. */
-  candidateCount?: number
-  /** Legacy lexical retrieval option. No longer used. */
-  includeKeyword?: boolean
-  /** Legacy reranker option. No longer used. */
-  rerankTopN?: number
-  /** User ID for memory isolation filtering. No longer used by retrieval. */
-  userId?: string
-}
-
-export interface RetrieveContextDebug {
-  subqueries: string[]
-  source_types: string[] | null
-  metadata_filter: Record<string, string> | null
-  candidate_count_before_rerank: number
-  reranked_count: number
-  fallback_used: boolean
-}
-
-export interface RetrieveContextResult {
-  chunks: RetrievedChunk[]
-  debug: RetrieveContextDebug
-}
-
-/** Intent -> legacy source types. null = all sources. */
-const INTENT_SOURCE_ROUTES: Record<IntentType, string[] | null> = {
-  product_recommendation: ["product_list", "book", "community_qa"],
-  ingredient_question: ["book", "qa"],
-  hair_care_advice: ["book", "transcript", "qa", "product_list", "community_qa"],
-  routine_help: ["book", "transcript", "qa", "product_list", "community_qa"],
-  diagnosis: ["book", "qa", "live_call", "community_qa", "product_list"],
-  general_chat: null,
-  followup: null,
-}
-
-export function computeChunkBoostedScore(
-  chunk: Pick<RetrievedChunk, "metadata" | "source_type" | "weighted_similarity" | "similarity">,
-  hairProfile: HairProfile | null | undefined,
-  shampooConcern?: string | null,
-): number {
-  let score = chunk.weighted_similarity ?? chunk.similarity ?? 0
-
-  if (hairProfile?.thickness && chunk.metadata?.thickness === hairProfile.thickness) {
-    score *= THICKNESS_MATCH_BOOST
-  }
-
-  if (shampooConcern && chunk.metadata?.concern === shampooConcern) {
-    score *= SHAMPOO_CONCERN_MATCH_BOOST
-  }
-
-  if (
-    chunk.source_type === "community_qa" &&
-    Array.isArray(chunk.metadata?.products_mentioned) &&
-    (chunk.metadata.products_mentioned as string[]).length > 0
-  ) {
-    score *= COMMUNITY_QA_PRODUCT_BOOST
-  }
-
-  return score
-}
-
-export async function retrieveContext(
-  query: string,
-  options: RetrieveOptions = {},
-): Promise<RetrieveContextResult> {
-  void query
-  void options.hairProfile
-  void options.shampooConcern
-  void options.count
-  void options.candidateCount
-  void options.includeKeyword
-  void options.rerankTopN
-  void options.userId
-
-  return {
-    chunks: [],
-    debug: {
-      subqueries: options.subqueries ?? [],
-      source_types: options.intent ? (INTENT_SOURCE_ROUTES[options.intent] ?? null) : null,
-      metadata_filter: options.metadataFilter ?? null,
-      candidate_count_before_rerank: 0,
-      reranked_count: 0,
-      fallback_used: false,
-    },
-  }
-}
