--- src/lib/rag/selection/types.ts
@@ -1,5 +1,0 @@
-import type { MatchedProduct } from "@/lib/product-matching/matcher"
-
-export interface SelectionResult {
-  products: MatchedProduct[]
-}
