--- tests/routine-planner.spec.ts
@@ -1,1666 +1,0 @@
-import { expect, test } from "@playwright/test"
-import {
-  activateRoutineTopics,
-  buildRoutinePlan,
-  buildRoutineRetrievalSubqueries,
-  deriveRoutineContext,
-  detectStylingProductKind,
-  getRoutineAutofillSlots,
-  projectRoutinePlanForLayer,
-} from "../src/lib/routines/planner"
-import { applyConversationStateToClassification } from "../src/lib/chat-runtime/conversation-state"
-import { evaluateRoute } from "../src/lib/rag/router"
-import { buildSystemPrompt } from "../src/lib/rag/synthesizer"
-import type {
-  ClassificationResult,
-  ContentChunk,
-  ConversationState,
-  HairProfile,
-} from "../src/lib/types"
-
-function createProfile(overrides: Partial<HairProfile> = {}): HairProfile {
-  return {
-    id: "profile-1",
-    user_id: "user-1",
-    hair_texture: "wavy",
-    thickness: "fine",
-    density: "medium",
-    concerns: [],
-    products_used: null,
-    wash_frequency: "every_2_3_days",
-    heat_styling: "never",
-    styling_tools: [],
-    goals: [],
-    cuticle_condition: "smooth",
-    protein_moisture_balance: "stretches_bounces",
-    scalp_type: "balanced",
-    scalp_condition: null,
-    chemical_treatment: ["natural"],
-    desired_volume: "balanced",
-    routine_preference: "advanced",
-    current_routine_products: ["shampoo", "conditioner"],
-    towel_material: null,
-    towel_technique: null,
-    drying_method: "air_dry",
-    brush_type: null,
-    night_protection: [],
-    uses_heat_protection: false,
-    additional_notes: null,
-    conversation_memory: null,
-    created_at: "2026-04-09T00:00:00.000Z",
-    updated_at: "2026-04-09T00:00:00.000Z",
-    ...overrides,
-  }
-}
-
-function createRoutineClassification(
-  overrides: Partial<ClassificationResult> = {},
-): ClassificationResult {
-  return {
-    intent: "routine_help",
-    product_category: "routine",
-    complexity: "multi_constraint",
-    needs_clarification: false,
-    retrieval_mode: "hybrid",
-    normalized_filters: {
-      problem: null,
-      duration: null,
-      products_tried: null,
-      routine: null,
-      special_circumstances: null,
-    },
-    router_confidence: 0.92,
-    ...overrides,
-  }
-}
-
-function createChunk(overrides: Partial<ContentChunk> = {}): ContentChunk {
-  return {
-    id: "chunk-1",
-    source_type: "book",
-    source_name: "Routine Basics",
-    chunk_index: 0,
-    content: "Routine content",
-    token_count: 12,
-    metadata: {},
-    created_at: "2026-04-09T00:00:00.000Z",
-    ...overrides,
-  }
-}
-
-test.describe("Routine planner", () => {
-  test("wavy frizz routine stays structure-first and adds one lightweight finish slot", () => {
-    const profile = createProfile({
-      concerns: ["frizz"],
-      goals: ["less_frizz"],
-      current_routine_products: ["shampoo", "conditioner"],
-    })
-
-    const plan = buildRoutinePlan(profile, "Welche Routine empfiehlst du gegen Frizz?")
-    const activeTopics = plan.active_topics.map((topic) => topic.label)
-    const baseSlots = plan.sections.find((section) => section.phase === "base_wash")?.slots ?? []
-    const maintenanceSlots =
-      plan.sections.find((section) => section.phase === "maintenance")?.slots ?? []
-
-    expect(activeTopics).toContain("Locken & Wellen")
-    expect(activeTopics).toContain("Lockenrefresh")
-    expect(activeTopics).not.toContain("Tiefenreinigung")
-    expect(activeTopics).not.toContain("Hair Oiling")
-
-    expect(baseSlots.find((slot) => slot.label === "Shampoo")?.action).toBe("keep")
-    expect(baseSlots.find((slot) => slot.label === "Conditioner")?.action).toBe("keep")
-
-    const leaveInSlot = maintenanceSlots.find((slot) => slot.label === "Leave-in / Finish")
-    expect(leaveInSlot?.action).toBe("add")
-    expect(leaveInSlot?.product_linkable).toBe(true)
-  })
-
-  test("curly hair with between-wash days activates routine locken and lockenrefresh", () => {
-    const topics = activateRoutineTopics(
-      createProfile({
-        hair_texture: "curly",
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Meine Locken sehen am naechsten Tag schnell platt aus.",
-    )
-
-    expect(topics.map((topic) => topic.id)).toEqual(
-      expect.arrayContaining(["routine_locken", "lockenrefresh"]),
-    )
-  })
-
-  test("oily scalp activates a dedicated scalp clarify slot without forcing hair reset", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        scalp_type: "oily",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Meine Kopfhaut fettet schnell nach.",
-    )
-
-    const topicIds = plan.active_topics.map((topic) => topic.id)
-    const scalpClarifySlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-scalp-clarify")
-    const hairResetSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-hair-reset")
-
-    expect(topicIds).toContain("tiefenreinigung")
-    expect(scalpClarifySlot?.label).toBe("Kopfhaut-Tiefenreinigung")
-    expect(hairResetSlot).toBeUndefined()
-  })
-
-  test("explicit requested category forces a leave-in slot even when profile heuristics suppress it", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "straight",
-        thickness: "fine",
-        concerns: ["oily_scalp"],
-        goals: ["shine"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Bau ein Leave-in in meine Routine ein.",
-      { forceRequestedCategory: "leave_in" },
-    )
-
-    const slots = plan.sections.flatMap((section) => section.slots)
-    const leaveInSlot = slots.find((slot) => slot.category === "leave_in")
-
-    expect(leaveInSlot?.id).toBe("maintenance-leave-in")
-    expect(leaveInSlot?.action).toBe("add")
-    expect(leaveInSlot?.rationale.join(" ")).toMatch(/ausdruecklich|aufnehmen/i)
-    expect(leaveInSlot?.rationale.join(" ")).not.toMatch(/Routine nach dem Waschen runder/i)
-    expect(plan.layer_projections?.basics.visible_slot_ids).toEqual([
-      "base-shampoo",
-      "base-conditioner",
-      "maintenance-leave-in",
-    ])
-  })
-
-  test("volume goal alone does not activate tiefenreinigung", () => {
-    const topics = activateRoutineTopics(
-      createProfile({
-        scalp_type: "balanced",
-        goals: ["volume"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Ich will einfach nur mehr Volumen.",
-    )
-
-    expect(topics.map((topic) => topic.id)).not.toContain("tiefenreinigung")
-  })
-
-  test("hard water exposure activates hair reset and its aftercare guidance", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        scalp_type: "balanced",
-        current_routine_products: ["shampoo", "conditioner"],
-        products_used: "Ich habe hartes Wasser und meine Haare werden schnell stumpf.",
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const hairResetSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-hair-reset")
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("tiefenreinigung")
-    expect(hairResetSlot?.label).toBe("Haar-Reset / Tiefenreinigung")
-    expect(hairResetSlot?.rationale.some((line) => line.includes("Conditioner oder Maske"))).toBe(
-      true,
-    )
-  })
-
-  test("chlorine and product overload escalate to hard reset", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        current_routine_products: ["shampoo", "conditioner", "leave_in", "mask", "oil"],
-        products_used: "Ich schwimme oft im Chlorwasser, nutze Gel und viele Produkte in Rotation.",
-      }),
-      "Meine Haare fuehlen sich wachsig und beschwert an.",
-    )
-
-    const hairResetSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-hair-reset")
-    const hardResetSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-hard-reset")
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("tiefenreinigung")
-    expect(hairResetSlot).toBeDefined()
-    expect(hardResetSlot?.label).toBe("Hard Reset")
-    expect(hardResetSlot?.rationale.some((line) => line.includes("Conditioner oder Maske"))).toBe(
-      true,
-    )
-  })
-
-  test("sensitive scalp suppresses hard reset even with overload signals", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        scalp_condition: "dry_flakes",
-        current_routine_products: ["shampoo", "conditioner", "leave_in", "mask", "oil"],
-        products_used: "Ich schwimme oft im Chlorwasser und meine Haare fuehlen sich wachsig an.",
-      }),
-      "Meine Haare fuehlen sich beschwert an.",
-    )
-
-    const hairResetSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-hair-reset")
-    const hardResetSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-hard-reset")
-
-    expect(hairResetSlot).toBeDefined()
-    expect(hardResetSlot).toBeUndefined()
-  })
-
-  test("damage signals activate bond builder and upgrade the conditioner slot", () => {
-    const profile = createProfile({
-      concerns: ["hair_damage"],
-      cuticle_condition: "rough",
-      chemical_treatment: ["bleached"],
-      current_routine_products: ["shampoo", "conditioner"],
-      protein_moisture_balance: "stretches_stays",
-    })
-
-    const plan = buildRoutinePlan(profile, "Meine Haare sind blondiert und brechen schnell.")
-    const activeTopicIds = plan.active_topics.map((topic) => topic.id)
-    const conditionerSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.label === "Conditioner")
-
-    expect(activeTopicIds).toContain("bond_builder")
-    expect(conditionerSlot?.action).toBe("upgrade")
-  })
-
-  test("existing leave-in can be adjusted instead of added", () => {
-    const profile = createProfile({
-      concerns: ["frizz"],
-      current_routine_products: ["shampoo", "conditioner", "leave_in"],
-    })
-
-    const plan = buildRoutinePlan(profile, "Ich brauche eine Routine fuer weniger Frizz.", {
-      forceRequestedCategory: "leave_in",
-    })
-    const leaveInSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.label === "Leave-in / Finish")
-
-    expect(leaveInSlot?.action).toBe("adjust")
-    expect(leaveInSlot?.rationale.join(" ")).toMatch(/Routine nach dem Waschen runder/i)
-    expect(leaveInSlot?.rationale.join(" ")).not.toMatch(/ausdruecklich|aufnehmen/i)
-  })
-
-  test("dryness with low oil-risk activates hair oiling", () => {
-    const topics = activateRoutineTopics(
-      createProfile({
-        hair_texture: "straight",
-        thickness: "normal",
-        density: "low",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Ich suche eine Routine fuer trockene Laengen.",
-    )
-
-    expect(topics.map((topic) => topic.id)).toContain("hair_oiling")
-  })
-
-  test("oily scalp does not activate hair oiling", () => {
-    const profile = createProfile({
-      scalp_type: "oily",
-      concerns: ["oily_scalp"],
-      current_routine_products: ["shampoo", "conditioner"],
-    })
-
-    const topics = activateRoutineTopics(
-      profile,
-      "Meine Kopfhaut fettet schnell nach. Welche Routine passt?",
-    )
-    expect(topics.map((topic) => topic.id)).not.toContain("hair_oiling")
-
-    const plan = buildRoutinePlan(
-      profile,
-      "Meine Kopfhaut fettet schnell nach. Welche Routine passt?",
-    )
-    const oilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-oil")
-    expect(oilSlot).toBeUndefined()
-  })
-
-  test("healthy_scalp goal alone does not activate hair oiling", () => {
-    const topics = activateRoutineTopics(
-      createProfile({
-        scalp_type: "oily",
-        goals: ["healthy_scalp"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    expect(topics.map((topic) => topic.id)).not.toContain("hair_oiling")
-  })
-
-  test("dandruff does not auto-activate hair oiling or tiefenreinigung", () => {
-    const topics = activateRoutineTopics(
-      createProfile({
-        scalp_condition: "dandruff",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    expect(topics.map((topic) => topic.id)).not.toContain("hair_oiling")
-    expect(topics.map((topic) => topic.id)).not.toContain("tiefenreinigung")
-  })
-
-  test("active oiling slot includes wash-out rationale and essential oil caveat", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "straight",
-        thickness: "normal",
-        density: "low",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Ich suche eine Routine fuer trockene Laengen.",
-    )
-
-    const oilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-oil")
-
-    expect(oilSlot?.action).toBe("add")
-    expect(
-      oilSlot?.rationale.some((line) => line.includes("Shampoo zuerst auf trockenes Haar")),
-    ).toBe(true)
-    expect(oilSlot?.caveats).toContain(
-      "Aetherische Oele (z.B. Rosmarin, Teebaum) nie pur auftragen — immer mit einem Basisoel verduennen.",
-    )
-  })
-
-  test("avoid oiling slot does not include wash-out rationale or essential oil caveat", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        current_routine_products: ["shampoo", "conditioner", "oil"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const oilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-oil")
-
-    expect(oilSlot?.action).toBe("avoid")
-    expect(
-      oilSlot?.rationale.every((line) => !line.includes("Shampoo zuerst auf trockenes Haar")),
-    ).toBe(true)
-    expect(oilSlot?.caveats).not.toContain(
-      "Aetherische Oele (z.B. Rosmarin, Teebaum) nie pur auftragen — immer mit einem Basisoel verduennen.",
-    )
-  })
-
-  test("irritated scalp gets both irritation caveat and essential oil caveat on active oiling", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "straight",
-        thickness: "normal",
-        density: "low",
-        scalp_type: "dry",
-        scalp_condition: "irritated",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Ich suche eine Routine fuer trockene Laengen.",
-    )
-
-    const oilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-oil")
-
-    expect(oilSlot?.caveats).toHaveLength(2)
-    expect(oilSlot?.caveats).toContain(
-      "Bei stark gereizter Kopfhaut eher sanft bleiben und die Routine nicht ueberladen.",
-    )
-    expect(oilSlot?.caveats).toContain(
-      "Aetherische Oele (z.B. Rosmarin, Teebaum) nie pur auftragen — immer mit einem Basisoel verduennen.",
-    )
-  })
-
-  test("leave-in activates independently of hair oiling", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        concerns: ["frizz"],
-        goals: ["less_frizz"],
-        scalp_type: "balanced",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const leaveInSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.label === "Leave-in / Finish")
-
-    expect(leaveInSlot).toBeDefined()
-    expect(leaveInSlot?.action).toBe("add")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("hair_oiling")
-  })
-
-  test("straight dry profile activates CWC as optional wash-day variant", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "straight",
-        concerns: ["dryness"],
-        chemical_treatment: ["colored"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine hilft gegen trockene Laengen?",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("cwc")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("owc")
-
-    const cwcSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-cwc-technique")
-
-    expect(cwcSlot?.phase).toBe("base_wash")
-    expect(cwcSlot?.cadence).toBe("als Wash-Day-Variante bei Bedarf")
-    expect(cwcSlot?.rationale[0]).toContain("kein Pflichtschritt")
-    expect(cwcSlot?.rationale.some((line) => line.startsWith("1. Conditioner"))).toBe(true)
-  })
-
-  test("wavy mild need stays on CWC", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "wavy",
-        concerns: ["frizz"],
-        cuticle_condition: "smooth",
-        chemical_treatment: ["natural"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Meine welligen Haare haben Frizz.",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("cwc")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("owc")
-  })
-
-  test("wavy profile upgrades to OWC only with multiple stronger support signals", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "wavy",
-        concerns: ["dryness"],
-        cuticle_condition: "slightly_rough",
-        chemical_treatment: ["colored"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Meine welligen Haare sind trocken und strapaziert.",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("owc")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("cwc")
-  })
-
-  test("curly dry profile gets OWC and a dedicated base-wash oil slot", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        thickness: "normal",
-        density: "medium",
-        concerns: [],
-        cuticle_condition: "smooth",
-        brush_type: "paddle",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine passt zu meinen Locken, wenn sie sich an Waschtagen schnell verhaken?",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("owc")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("hair_oiling")
-
-    const owcOilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-owc-oil")
-    const occasionalOilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-oil")
-
-    expect(owcOilSlot?.phase).toBe("base_wash")
-    expect(owcOilSlot?.category).toBe("oil")
-    expect(owcOilSlot?.action).toBe("add")
-    expect(owcOilSlot?.product_linkable).toBe(true)
-    expect(occasionalOilSlot).toBeUndefined()
-  })
-
-  test("OWC is not proactively activated when oil-weight risk is present", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        thickness: "fine",
-        density: "high",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine passt zu meinen Locken?",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("owc")
-    expect(
-      plan.sections.flatMap((section) => section.slots).find((slot) => slot.id === "base-owc-oil"),
-    ).toBeUndefined()
-  })
-
-  test("explicit OWC ask with oil-weight risk keeps the slot but disables oil product matching", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        thickness: "fine",
-        density: "high",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Soll ich OWC testen?",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("owc")
-
-    const owcOilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-owc-oil")
-
-    expect(owcOilSlot?.action).toBe("add")
-    expect(owcOilSlot?.product_linkable).toBe(false)
-    expect(owcOilSlot?.caveats.some((line) => line.includes("feinem, dichtem Haar"))).toBe(true)
-  })
-
-  test("comparison requests set compare mode and still personalize to one variant", () => {
-    const profile = createProfile({
-      hair_texture: "curly",
-      concerns: ["dryness"],
-      cuticle_condition: "rough",
-      current_routine_products: ["shampoo", "conditioner"],
-    })
-
-    const plan = buildRoutinePlan(profile, "Was ist der Unterschied zwischen CWC und OWC?")
-    const prompt = buildSystemPrompt(
-      profile,
-      [createChunk()],
-      [],
-      "routine",
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      plan,
-      null,
-      undefined,
-    )
-    const subqueries = buildRoutineRetrievalSubqueries(
-      "Was ist der Unterschied zwischen CWC und OWC?",
-      plan,
-    )
-
-    expect(plan.compare_cwc_owc).toBe(true)
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("owc")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("cwc")
-    expect(subqueries).toContain("CWC")
-    expect(subqueries).toContain("OWC")
-    expect(prompt).toContain("Vergleichsmodus")
-    expect(prompt).toContain("kompakten nummerierten Wash-Day-Schritte")
-  })
-
-  test("explicit CWC on healthy profile uses educational mode", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "straight",
-        concerns: [],
-        goals: [],
-        cuticle_condition: "smooth",
-        chemical_treatment: ["natural"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Was ist die CWC Methode?",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("cwc")
-
-    const cwcSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-cwc-technique")
-
-    expect(cwcSlot?.cadence).toBeNull()
-    expect(cwcSlot?.caveats.some((line) => line.includes("eher optional"))).toBe(true)
-    expect(cwcSlot?.product_linkable).toBe(false)
-    expect(cwcSlot?.rationale.some((line) => line.startsWith("1."))).toBe(false)
-  })
-
-  test("comparison on healthy profile does not activate any wash protection topic", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        concerns: [],
-        goals: [],
-        cuticle_condition: "smooth",
-        chemical_treatment: ["natural"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Was ist der Unterschied zwischen CWC und OWC?",
-    )
-
-    expect(plan.compare_cwc_owc).toBe(true)
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("cwc")
-    expect(plan.active_topics.map((topic) => topic.id)).not.toContain("owc")
-
-    const cwcSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-cwc-technique")
-    const owcSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-owc-technique")
-
-    expect(cwcSlot).toBeUndefined()
-    expect(owcSlot).toBeUndefined()
-  })
-
-  test("OWC suppresses duplicate occasional-oil slot when hair_oiling also activates", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        thickness: "normal",
-        density: "medium",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        scalp_condition: "dry_flakes",
-        brush_type: "paddle",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine passt zu meinen Locken?",
-    )
-
-    expect(plan.active_topics.map((topic) => topic.id)).toContain("owc")
-
-    const owcOilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "base-owc-oil")
-    const occasionalOilSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "occasional-oil")
-
-    expect(owcOilSlot).toBeDefined()
-    expect(owcOilSlot?.phase).toBe("base_wash")
-    expect(occasionalOilSlot).toBeUndefined()
-  })
-
-  test("unnecessary mask steps can still be deprioritized", () => {
-    const maskPlan = buildRoutinePlan(
-      createProfile({
-        current_routine_products: ["shampoo", "conditioner", "mask"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-    const maskSlot = maskPlan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.label === "Maske / Kur")
-
-    expect(maskSlot?.action).toBe("avoid")
-  })
-
-  test("router clarifies when the routine frame is still missing", () => {
-    const routerDecision = evaluateRoute(
-      createRoutineClassification({ router_confidence: 0.91 }),
-      [],
-      createProfile({
-        hair_texture: null,
-        concerns: [],
-        goals: [],
-        wash_frequency: null,
-        scalp_type: null,
-        current_routine_products: [],
-      }),
-      [],
-      "Welche Routine passt zu mir?",
-    )
-
-    expect(routerDecision.response_mode).toBe("clarify_only")
-    expect(routerDecision.policy_overrides).toContain("missing_routine_frame")
-  })
-
-  test("router still clarifies when organizer and cadence exist but inventory is missing", () => {
-    const routerDecision = evaluateRoute(
-      createRoutineClassification(),
-      [],
-      createProfile({
-        concerns: ["frizz"],
-        goals: ["less_frizz"],
-        wash_frequency: "every_2_3_days",
-        current_routine_products: [],
-        products_used: null,
-      }),
-      [],
-      "Welche Routine passt zu mir?",
-    )
-
-    expect(routerDecision.response_mode).toBe("clarify_only")
-    expect(routerDecision.policy_overrides).toContain("missing_routine_frame")
-  })
-
-  test("router can proceed once organizer, cadence, and inventory are available from the profile", () => {
-    const routerDecision = evaluateRoute(
-      createRoutineClassification(),
-      [],
-      createProfile({
-        concerns: ["frizz"],
-        goals: ["less_frizz"],
-        wash_frequency: "every_2_3_days",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      [],
-      "Welche Routine passt zu mir?",
-    )
-
-    expect(routerDecision.response_mode).not.toBe("clarify_only")
-    expect(routerDecision.policy_overrides).not.toContain("missing_routine_frame")
-  })
-
-  test("router proceeds when pending routine state receives the missing routine frame", () => {
-    const previousState = {
-      version: 1 as const,
-      active_topic: "routine" as const,
-      routine_layer: "basics" as const,
-      pending_offer: "routine_goals_or_problems" as const,
-      answered_slots: [],
-      last_assistant_action: "asked_routine_basics",
-      last_product_category: null,
-    }
-
-    const corrected = applyConversationStateToClassification({
-      state: previousState,
-      classification: createRoutineClassification({
-        intent: "followup",
-        product_category: null,
-        normalized_filters: {
-          problem: "trockene Spitzen",
-          duration: null,
-          products_tried: "Shampoo und Conditioner",
-          routine: "alle 3 Tage",
-          special_circumstances: null,
-        },
-        router_confidence: 0.58,
-      }),
-      userMessage: "Alle 3 Tage, Shampoo und Conditioner. Meine Spitzen sind trocken.",
-    })
-
-    const routerDecision = evaluateRoute(
-      corrected.classification,
-      [],
-      createProfile({
-        concerns: ["dryness"],
-        goals: ["less_frizz"],
-        wash_frequency: "every_2_3_days",
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      [],
-      "Alle 3 Tage, Shampoo und Conditioner. Meine Spitzen sind trocken.",
-    )
-
-    expect(corrected.override).toBe("conversation_state_pending_routine_answer")
-    expect(routerDecision.response_mode).not.toBe("clarify_only")
-    expect(routerDecision.policy_overrides).not.toContain("missing_routine_frame")
-  })
-
-  test("retrieval hints keep exact topic names and autofill only add or upgrade slots", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        concerns: ["frizz"],
-        goals: ["less_frizz"],
-        current_routine_products: ["shampoo", "conditioner"],
-      }),
-      "Welche Routine empfiehlst du fuer welliges Haar mit Frizz?",
-    )
-
-    const subqueries = buildRoutineRetrievalSubqueries(
-      "Welche Routine empfiehlst du fuer welliges Haar mit Frizz?",
-      plan,
-    )
-    const autofillSlots = getRoutineAutofillSlots(plan)
-
-    expect(subqueries.some((query) => query.includes("Locken & Wellen"))).toBe(true)
-    expect(autofillSlots.length).toBeGreaterThan(0)
-    expect(autofillSlots.every((slot) => slot.action === "add" || slot.action === "upgrade")).toBe(
-      true,
-    )
-  })
-
-  test("routine context keeps organizer and cadence signals independent from routine_preference", () => {
-    const context = deriveRoutineContext(
-      createProfile({
-        concerns: ["frizz"],
-        goals: ["less_frizz"],
-        routine_preference: "minimal",
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    expect(context.organizer_complete).toBe(true)
-    expect(context.cadence_complete).toBe(true)
-    expect(context.primary_focuses.map((focus) => focus.label)).not.toContain("minimal")
-  })
-
-  test.describe("detectStylingProductKind", () => {
-    test("detects Gel from free text", () => {
-      expect(detectStylingProductKind("Balea Styling Gel")).toBe("Gel")
-    })
-
-    test("detects Mousse/Schaum", () => {
-      expect(detectStylingProductKind("Ich nutze einen Locken-Schaum von Schwarzkopf")).toBe(
-        "Mousse",
-      )
-      expect(detectStylingProductKind("Mousse von Cantu")).toBe("Mousse")
-    })
-
-    test("detects Lockencreme before generic Creme", () => {
-      expect(detectStylingProductKind("Lockencreme von Cantu und Gel")).toBe("Lockencreme")
-      expect(detectStylingProductKind("Curl Cream plus Gel")).toBe("Lockencreme")
-    })
-
-    test("detects generic Stylingcreme", () => {
-      expect(detectStylingProductKind("eine Creme fuer die Haare")).toBe("Stylingcreme")
-    })
-
-    test("returns null for empty or no-match input", () => {
-      expect(detectStylingProductKind(null)).toBeNull()
-      expect(detectStylingProductKind("")).toBeNull()
-      expect(detectStylingProductKind("ich benutze nichts besonderes")).toBeNull()
-    })
-
-    test("handles mixed casing and diacritics", () => {
-      expect(detectStylingProductKind("STYLING GEL von Wella")).toBe("Gel")
-      expect(detectStylingProductKind("Schäum-Mousse")).toBe("Mousse")
-    })
-
-    test("matches German compound words", () => {
-      expect(detectStylingProductKind("Stylinggel von dm")).toBe("Gel")
-      expect(detectStylingProductKind("Lockenschaum")).toBe("Mousse")
-    })
-
-    test("does not match gel inside gelb", () => {
-      expect(detectStylingProductKind("gelbliches Serum")).toBeNull()
-    })
-  })
-
-  test("refresh slot echoes product kind from products_used", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        products_used: "Balea Styling Gel",
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(refreshSlot).toBeDefined()
-    expect(
-      refreshSlot?.rationale.some((line) => line.includes("dein Gel vom letzten Waschtag")),
-    ).toBe(true)
-  })
-
-  test("refresh slot uses generic product echo when products_used has no styling match", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        products_used: null,
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(
-      refreshSlot?.rationale.some((line) =>
-        line.includes("dasselbe Styling-Produkt vom letzten Waschtag"),
-      ),
-    ).toBe(true)
-  })
-
-  test("refresh slot adds leave-in cross-reference for dry/damaged hair", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        concerns: ["dryness"],
-        cuticle_condition: "rough",
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(refreshSlot?.rationale.some((line) => line.includes("Leave-In"))).toBe(true)
-    expect(refreshSlot?.rationale.some((line) => line.includes("trainiert langfristig"))).toBe(true)
-  })
-
-  test("refresh slot adds fine-hair caveat when thickness is fine and dryness signals present", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        thickness: "fine",
-        concerns: ["dryness"],
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(refreshSlot?.caveats.some((line) => line.includes("feinem Haar"))).toBe(true)
-  })
-
-  test("refresh slot has no leave-in cross-reference when no dryness signals", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        concerns: [],
-        goals: [],
-        cuticle_condition: "smooth",
-        chemical_treatment: ["natural"],
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(refreshSlot?.rationale.every((line) => !line.includes("Leave-In"))).toBe(true)
-    expect(refreshSlot?.caveats).toHaveLength(0)
-  })
-
-  test("refresh slot includes duration in cadence", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(refreshSlot?.cadence).toContain("ca. 10 Min.")
-  })
-
-  test("chemical treatment alone triggers leave-in cross-reference on refresh", () => {
-    const plan = buildRoutinePlan(
-      createProfile({
-        hair_texture: "curly",
-        thickness: "normal",
-        chemical_treatment: ["colored"],
-        cuticle_condition: "smooth",
-        concerns: [],
-        goals: [],
-        current_routine_products: ["shampoo", "conditioner", "leave_in"],
-      }),
-      "Welche Routine passt zu mir?",
-    )
-
-    const refreshSlot = plan.sections
-      .flatMap((section) => section.slots)
-      .find((slot) => slot.id === "maintenance-refresh")
-
-    expect(refreshSlot?.rationale.some((line) => line.includes("Leave-In"))).toBe(true)
-    expect(refreshSlot?.caveats).toHaveLength(0)
-  })
-
-  test("routine system prompt includes the plan, softer avoid wording, and reasoning-first product instructions", () => {
-    const profile = createProfile({
-      current_routine_products: ["shampoo", "conditioner", "mask"],
-      routine_preference: "advanced",
-    })
-    const routinePlan = buildRoutinePlan(profile, "Welche Routine passt zu mir?")
-
-    const prompt = buildSystemPrompt(
-      profile,
-      [createChunk()],
-      [],
-      "routine",
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      routinePlan,
-      null,
-      undefined,
-    )
-
-    expect(prompt).toContain("Routine-Plan:")
-    expect(prompt).toContain("Maske / Kur: gerade eher weniger geeignet")
-    expect(prompt).toContain("Begruende pro relevantem Slot erst kurz den Fit zum Profil")
-    expect(prompt).toContain("ordne sie direkt dem gerade erklaerten Slot")
-    expect(prompt).toContain(
-      "unterscheide sauber zwischen Kopfhaut-Tiefenreinigung, Haar-Reset und Hard Reset",
-    )
-    expect(prompt).toContain("Tiefenreinigung nie als universellen Pflichtschritt")
-    expect(prompt).not.toContain("Routine-Detailgrad")
-  })
-
-  test("system prompt includes compact conversation state when an active topic exists", () => {
-    const conversationState: ConversationState = {
-      version: 1,
-      active_topic: "routine",
-      routine_layer: "basics",
-      pending_offer: "routine_goals_or_problems",
-      answered_slots: ["routine_frame"],
-      last_assistant_action: "asked_routine_basics",
-      last_product_category: null,
-    }
-
-    const prompt = buildSystemPrompt(
-      createProfile(),
-      [createChunk()],
-      [],
-      "routine",
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      null,
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      conversationState,
-    )
-
-    expect(prompt).toContain("<conversation_state>")
-    expect(prompt).toContain("Aktives Thema: routine")
-    expect(prompt).toContain("Routine-Ebene: basics")
-    expect(prompt).toContain("Beantwortete Slots: routine_frame")
-  })
-
-  test("system prompt omits conversation state when no active topic exists", () => {
-    const prompt = buildSystemPrompt(
-      createProfile(),
-      [createChunk()],
-      [],
-      "routine",
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      null,
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      {
-        version: 1,
-        active_topic: null,
-        routine_layer: null,
-        pending_offer: null,
-        answered_slots: [],
-        last_assistant_action: null,
-        last_product_category: null,
-      },
-    )
-
-    expect(prompt).not.toContain("<conversation_state>")
-  })
-
-  test("product prompt uses not-recommended wording for explicit support categories", () => {
-    const prompt = buildSystemPrompt(
-      createProfile(),
-      [createChunk()],
-      [],
-      "bondbuilder",
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      undefined,
-      null,
-      undefined,
-      undefined,
-      undefined,
-      {
-        category: "bondbuilder",
-        relevant: false,
-        action: null,
-        planReasonCodes: [],
-        currentInventory: null,
-        targetProfile: null,
-        notes: ["bond_builder_low_relevance_currently"],
-      },
-    )
-
-    expect(prompt).toContain("aktuell kein Bondbuilder empfohlen ist")
-    expect(prompt).toContain("nicht als Katalog-No-Match")
-    expect(prompt).not.toContain("Keine passenden Produkte in der Datenbank gefunden")
-  })
-
-  test.describe("Bond builder logic", () => {
-    test("colored alone does NOT activate bond builder", () => {
-      const topics = activateRoutineTopics(
-        createProfile({
-          chemical_treatment: ["colored"],
-          cuticle_condition: "smooth",
-          concerns: [],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      expect(topics.map((topic) => topic.id)).not.toContain("bond_builder")
-    })
-
-    test("colored + rough cuticle activates bond builder", () => {
-      const topics = activateRoutineTopics(
-        createProfile({
-          chemical_treatment: ["colored"],
-          cuticle_condition: "rough",
-          concerns: [],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      expect(topics.map((topic) => topic.id)).toContain("bond_builder")
-    })
-
-    test("heat damage without protection activates bond builder", () => {
-      const topics = activateRoutineTopics(
-        createProfile({
-          heat_styling: "daily",
-          uses_heat_protection: false,
-          cuticle_condition: "smooth",
-          concerns: [],
-          chemical_treatment: ["natural"],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      expect(topics.map((topic) => topic.id)).toContain("bond_builder")
-    })
-
-    test("explicit request without damage signals uses educational mode", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          cuticle_condition: "smooth",
-          chemical_treatment: ["natural"],
-          concerns: [],
-          heat_styling: "never",
-        }),
-        "Was ist ein Bond Builder?",
-      )
-
-      const bondSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-      const topicIds = plan.active_topics.map((topic) => topic.id)
-
-      expect(bondSlot).toBeDefined()
-      expect(bondSlot?.caveats.some((line) => line.includes("eher optional"))).toBe(true)
-      expect(bondSlot?.cadence).toBeNull()
-      expect(bondSlot?.rationale.some((line) => line.includes("optionaler Baustein"))).toBe(true)
-      expect(topicIds).toContain("bond_builder")
-      expect(topicIds).not.toContain("tiefenreinigung")
-    })
-
-    test("snaps = severe tier with pro note", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          protein_moisture_balance: "snaps",
-          chemical_treatment: ["bleached"],
-          cuticle_condition: "rough",
-          concerns: ["hair_damage"],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const bondSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(
-        bondSlot?.rationale.some((line) => line.includes("Kombination aus K18 und Olaplex")),
-      ).toBe(true)
-      expect(
-        bondSlot?.caveats.some((line) => line.includes("professionelles Beratungsgespraech")),
-      ).toBe(true)
-    })
-
-    test("moderate + chemical treatment leans Olaplex", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          chemical_treatment: ["colored"],
-          cuticle_condition: "rough",
-          concerns: [],
-          protein_moisture_balance: "stretches_bounces",
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const bondSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(bondSlot?.rationale.some((line) => line.includes("Olaplex (Querverbindungen)"))).toBe(
-        true,
-      )
-    })
-
-    test("moderate + no chemical treatment leans K18", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          chemical_treatment: ["natural"],
-          cuticle_condition: "rough",
-          concerns: ["hair_damage"],
-          protein_moisture_balance: "stretches_bounces",
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const bondSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(bondSlot?.rationale.some((line) => line.includes("K18 (Laengsverbindungen)"))).toBe(
-        true,
-      )
-    })
-
-    test("bond builder co-activates tiefenreinigung", () => {
-      const profile = createProfile({
-        cuticle_condition: "rough",
-        concerns: ["hair_damage"],
-        chemical_treatment: ["bleached"],
-      })
-      const topics = activateRoutineTopics(profile, "Welche Routine passt zu mir?")
-      const plan = buildRoutinePlan(profile, "Welche Routine passt zu mir?")
-
-      const topicIds = topics.map((topic) => topic.id)
-      expect(topicIds).toContain("bond_builder")
-      expect(topicIds).toContain("tiefenreinigung")
-
-      const tiefenreinigung = topics.find((topic) => topic.id === "tiefenreinigung")
-      expect(tiefenreinigung?.reason).toContain("Rueckstaende")
-
-      const hairResetSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-hair-reset")
-      expect(hairResetSlot?.rationale.some((line) => line.includes("Bond Builder"))).toBe(true)
-    })
-
-    test("repair fatigue caveat is always present on bond builder slot", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          cuticle_condition: "rough",
-          concerns: ["hair_damage"],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const bondSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(bondSlot?.caveats.some((line) => line.includes("steif und sproede"))).toBe(true)
-    })
-
-    test("protein interaction caveat varies by balance", () => {
-      const stretchesPlan = buildRoutinePlan(
-        createProfile({
-          cuticle_condition: "rough",
-          concerns: ["hair_damage"],
-          protein_moisture_balance: "stretches_stays",
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const stretchesSlot = stretchesPlan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(stretchesSlot?.caveats.some((line) => line.includes("parallel laufen"))).toBe(true)
-
-      const balancedPlan = buildRoutinePlan(
-        createProfile({
-          cuticle_condition: "rough",
-          concerns: ["hair_damage"],
-          protein_moisture_balance: "stretches_bounces",
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const balancedSlot = balancedPlan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(balancedSlot?.caveats.some((line) => line.includes("Feuchtigkeit reicht"))).toBe(true)
-    })
-
-    test("system prompt includes bond builder synth rules when bond_builder is active", () => {
-      const profile = createProfile({
-        cuticle_condition: "rough",
-        concerns: ["hair_damage"],
-        chemical_treatment: ["bleached"],
-      })
-      const routinePlan = buildRoutinePlan(profile, "Welche Routine passt zu mir?")
-
-      const prompt = buildSystemPrompt(
-        profile,
-        [createChunk()],
-        [],
-        "routine",
-        undefined,
-        undefined,
-        undefined,
-        undefined,
-        routinePlan,
-        null,
-        undefined,
-      )
-
-      expect(prompt).toContain("nachgewiesener Bond-Technologie")
-      expect(prompt).toContain("Laengs- und Querverbindungen")
-    })
-
-    test("snaps alone activates bond builder via damage signals", () => {
-      const topics = activateRoutineTopics(
-        createProfile({
-          protein_moisture_balance: "snaps",
-          cuticle_condition: "smooth",
-          chemical_treatment: ["natural"],
-          concerns: [],
-          heat_styling: "never",
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      expect(topics.map((topic) => topic.id)).toContain("bond_builder")
-    })
-
-    test("usesBondBuilder flag sets slot action to adjust", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          cuticle_condition: "rough",
-          concerns: ["hair_damage"],
-        }),
-        "Welche Routine passt zu mir?",
-        { usesBondBuilder: true },
-      )
-
-      const bondSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "occasional-bond-builder")
-
-      expect(bondSlot?.action).toBe("adjust")
-    })
-  })
-
-  test.describe("Brush tools logic", () => {
-    test("explicit brush question activates brush_tools", () => {
-      const topics = activateRoutineTopics(
-        createProfile(),
-        "Welche Buerste ist fuer nasses Entwirren am besten?",
-      )
-
-      expect(topics.map((topic) => topic.id)).toContain("brush_tools")
-    })
-
-    test("rough brushing proactively adds a brush tools slot", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          brush_type: "paddle",
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const topicIds = plan.active_topics.map((topic) => topic.id)
-      const brushSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "maintenance-brush-tools")
-
-      expect(topicIds).toContain("brush_tools")
-      expect(brushSlot?.action).toBe("adjust")
-      expect(brushSlot?.rationale.some((line) => line.includes("von unten nach oben"))).toBe(true)
-      expect(brushSlot?.rationale.some((line) => line.includes("Paddle- und Rundbuersten"))).toBe(
-        true,
-      )
-    })
-
-    test("scalp-sensitive profiles get gentle scalp tool caveats", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          scalp_condition: "irritated",
-          concerns: ["hair_loss"],
-        }),
-        "Welches Tool passt fuer Kopfhautserum und Scalp Brush?",
-      )
-
-      const brushSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "maintenance-brush-tools")
-
-      expect(brushSlot?.rationale.some((line) => line.includes("Applikatorflasche"))).toBe(true)
-      expect(
-        brushSlot?.caveats.some((line) => line.includes("gereizter oder schuppiger Kopfhaut")),
-      ).toBe(true)
-      expect(brushSlot?.caveats.some((line) => line.includes("Haarausfall oder Ausduennung"))).toBe(
-        true,
-      )
-    })
-
-    test("curl refresh tool questions surface the spray-bottle hint", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          hair_texture: "curly",
-          wash_frequency: "every_2_3_days",
-        }),
-        "Brauche ich eine Spruehflasche fuer meinen Lockenrefresh?",
-      )
-
-      const brushSlot = plan.sections
-        .flatMap((section) => section.slots)
-        .find((slot) => slot.id === "maintenance-brush-tools")
-
-      expect(plan.active_topics.map((topic) => topic.id)).toContain("brush_tools")
-      expect(brushSlot?.rationale.some((line) => line.includes("Spruehflasche"))).toBe(true)
-    })
-  })
-
-  test.describe("Layered routine projections", () => {
-    test("basics projection keeps shampoo, conditioner, and exactly one separate priority lever when available", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          concerns: ["frizz"],
-          goals: ["less_frizz"],
-          current_routine_products: ["shampoo", "conditioner"],
-        }),
-        "Welche Routine empfiehlst du gegen Frizz?",
-      )
-
-      const projection = projectRoutinePlanForLayer(plan, "basics")
-
-      expect(plan.priority_lever?.slot_id).toBe("maintenance-leave-in")
-      expect(projection.priority_lever?.slot_id).toBe("maintenance-leave-in")
-      expect(projection.visible_slot_ids).toEqual([
-        "base-shampoo",
-        "base-conditioner",
-        "maintenance-leave-in",
-      ])
-    })
-
-    test("severe active damage selects care-product-first and starts with the conditioner upgrade", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          concerns: ["breakage", "hair_damage", "split_ends"],
-          cuticle_condition: "rough",
-          chemical_treatment: ["bleached"],
-          protein_moisture_balance: "snaps",
-          current_routine_products: ["shampoo", "conditioner"],
-        }),
-        "Meine Haare sind blondiert, rau und brechen aktiv ab.",
-      )
-
-      expect(plan.priority_lever).toMatchObject({
-        id: "care-product-first",
-        source: "care_risk",
-        slot_id: "base-conditioner",
-      })
-      expect(plan.priority_lever?.supporting_slot_ids).toEqual(
-        expect.arrayContaining([
-          "maintenance-leave-in",
-          "occasional-mask",
-          "occasional-bond-builder",
-        ]),
-      )
-    })
-
-    test("severe active damage still selects care-product-first when reset signal is weak oil usage", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          concerns: ["breakage", "hair_damage", "split_ends"],
-          cuticle_condition: "rough",
-          chemical_treatment: ["bleached"],
-          protein_moisture_balance: "snaps",
-          current_routine_products: ["shampoo", "conditioner", "oil"],
-        }),
-        "Meine Haare sind blondiert, rau und brechen aktiv ab.",
-      )
-
-      expect(plan.priority_lever).toMatchObject({
-        id: "care-product-first",
-        source: "care_risk",
-        slot_id: "base-conditioner",
-      })
-    })
-
-    test("strong reset blockage can preempt cosmetic goals", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          goals: ["shine", "curl_definition"],
-          products_used:
-            "Hartes Wasser, Gel und viele Produkte in Rotation; die Haare fuehlen sich wachsig an und nichts zieht mehr ein.",
-          current_routine_products: ["shampoo", "conditioner", "leave_in", "mask", "oil"],
-        }),
-        "Ich will mehr Glanz, aber meine Haare sind wachsig und beschwert.",
-      )
-
-      expect(plan.priority_lever).toMatchObject({
-        id: "reset-blockage",
-        source: "care_risk",
-        slot_id: "occasional-hair-reset",
-      })
-    })
-
-    test("stated goal wins when no stronger care risk is present", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          hair_texture: "curly",
-          concerns: [],
-          goals: ["curl_definition"],
-          cuticle_condition: "smooth",
-          chemical_treatment: ["natural"],
-          current_routine_products: ["shampoo", "conditioner"],
-        }),
-        "Ich moechte vor allem mehr Definition.",
-      )
-
-      expect(plan.priority_lever).toMatchObject({
-        id: "stated-goal",
-        source: "stated_goal",
-        slot_id: "maintenance-leave-in",
-      })
-    })
-
-    test("goals and problems projections expose only the top two to three relevant levers", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          hair_texture: "curly",
-          concerns: ["dryness", "frizz", "tangling", "hair_damage"],
-          goals: ["less_frizz", "curl_definition", "moisture", "healthier_hair"],
-          cuticle_condition: "rough",
-          chemical_treatment: ["colored"],
-          current_routine_products: ["shampoo", "conditioner"],
-        }),
-        "Welche Routine passt fuer definierte, weniger trockene Locken?",
-      )
-
-      const goalsProjection = projectRoutinePlanForLayer(plan, "goals")
-      const problemsProjection = projectRoutinePlanForLayer(plan, "problems")
-
-      expect(goalsProjection.visible_slot_ids).toHaveLength(3)
-      expect(problemsProjection.visible_slot_ids).toHaveLength(3)
-      expect(goalsProjection.visible_slot_ids).toEqual(
-        expect.arrayContaining(["maintenance-leave-in", "occasional-mask"]),
-      )
-      expect(problemsProjection.visible_slot_ids).toEqual(
-        expect.arrayContaining(["maintenance-leave-in", "occasional-mask"]),
-      )
-      expect(goalsProjection.visible_slot_ids).not.toContain("base-shampoo")
-      expect(goalsProjection.visible_slot_ids).not.toContain("base-conditioner")
-      expect(problemsProjection.visible_slot_ids).not.toContain("base-shampoo")
-      expect(problemsProjection.visible_slot_ids).not.toContain("base-conditioner")
-    })
-
-    test("deep dive projection focuses the requested category follow-up", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          concerns: ["frizz"],
-          goals: ["less_frizz"],
-          current_routine_products: ["shampoo", "conditioner"],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const projection = projectRoutinePlanForLayer(plan, "deep_dive", {
-        requestedCategory: "leave_in",
-      })
-
-      expect(projection.visible_slot_ids).toEqual(["maintenance-leave-in"])
-      expect(projection.requested_category).toBe("leave_in")
-    })
-
-    test("deep dive projection can focus base wash categories", () => {
-      const plan = buildRoutinePlan(
-        createProfile({
-          concerns: ["hair_damage"],
-          current_routine_products: ["shampoo", "conditioner"],
-        }),
-        "Welche Routine passt zu mir?",
-      )
-
-      const projection = projectRoutinePlanForLayer(plan, "deep_dive", {
-        requestedCategory: "conditioner",
-      })
-
-      expect(projection.visible_slot_ids).toEqual(["base-conditioner"])
-      expect(projection.requested_category).toBe("conditioner")
-    })
-  })
-})
